Task4 Внесение изменений

+++ src/com/max/idea/Task4.java
@@ -4,11 +4,11 @@ public class Task4 {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
         System.out.println("Введите число x: ");
-        double x = in.nextInt();
+        double x = in.nextDouble();
         System.out.println("Введите число y: ");
-        double y = in.nextInt();
+        double y = in.nextDouble();
         System.out.println("Введите число z: ");
-        double z = in.nextInt();
+        double z = in.nextDouble();
         double avr=(x+y+z)/3;
         System.out.printf("Ваши числа: x=%.2f, y=%.2f, z=%.2f \n", x,y,z);
         System.out.printf("Среднее арифметическое=%.2f \n", avr);
